Remove asset pipeline
Since we are not expected to style the app or perform some
clientside behavior we should remove delivering assets.
The letter-opener should be based on the web gem, because
in a docker environment the pure gem tries to open a web
browser in the container itself which isn't avaliable.
The web variant has the interface integrated in the
Rails app itself with its own shipped UI.