Introduce geocoding address
The users address provided in the registration process
is expected to be geocoded. Therefore the gem geocoder
should be added and wired with the public Nominatim API.
Turned out the longitude and latitude datapoints have a
scale of 7 digits. However the precision of 10 still is
considered sufficient.
Even though the geocoder gem already abstracts a lot of
API handling away we still should go with a dedicated
service object that is responsible for caching the
results as a instantiated object and mapping the datapoints
accordingly. Some datapoints like `state` have to be handled
different, because some addresses so not have a dedicated state
key. For instance in case of cities like Berlin or Hamburg being
a Bundesland too.