Assert server error handling
In case of exceptions with the 3rd party API it should be
catched properly. Turns out `Geocoder` already does the
job and we could rely on it.
However the default exception handling could switched off
and a custom exception handling could be hooked in. It may
make sense to follow that approach in order to have dedicated
outcomes (e.g. a special message in case of timeouts or such).